requirements.txt
openai
python-dotenv
tiktoken
pydantic
streamlit